backend/requirements.txt
# Core API
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9

# Database (SQLite for dev — swap for psycopg2-binary in production with PostgreSQL)
sqlalchemy==2.0.29
alembic==1.13.1
aiosqlite==0.20.0

# Data Processing
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1

# Environment
python-dotenv==1.0.1
pydantic-settings==2.2.1
pydantic[email]==2.6.4

# Serialisation
joblib==1.4.0
orjson==3.9.15

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6
